Add Section 2 standing rules, update template
Update SectionList component. Continue adding information from Camper's Guide PDF to sections constant.

Update SectionOverview.vue component template to conditionally render a list if it is present in the object, so we don't have to make a single line html str in the 'part.body' for parts with many points.